+++ src/libtrx/include/libtrx/game/game_string.def
@@ -73,6 +73,7 @@ GS_DEFINE(OSD_DOOR_OPEN, "Open Sesame!")
 GS_DEFINE(OSD_DOOR_CLOSE, "Close Sesame!")
 GS_DEFINE(OSD_DOOR_OPEN_FAIL, "No doors in Lara's proximity")
 GS_DEFINE(OSD_FLY_MODE_ON, "Fly mode enabled")
+GS_DEFINE(OSD_FLY_MODE_OFF, "Fly mode disabled")
 GS_DEFINE(CONTROLS_DEFAULT_KEYS, "Default Keys")
 GS_DEFINE(CONTROLS_CUSTOM_1, "User Keys 1")
 GS_DEFINE(CONTROLS_CUSTOM_2, "User Keys 2")
