+++ src/tr1/game/game_string.def
@@ -19,7 +19,6 @@ GS_DEFINE(STATS_BONUS_STATISTICS, "Bonus Statistics")
 GS_DEFINE(STATS_AMMO, "AMMO HITS/USED")
 GS_DEFINE(STATS_DISTANCE_TRAVELLED, "DISTANCE TRAVELLED")
 GS_DEFINE(STATS_MEDIPACKS_USED, "HEALTH PACKS USED")
-GS_DEFINE(OSD_FLY_MODE_OFF, "Fly mode disabled")
 GS_DEFINE(OSD_TEXTURE_FILTER_SET, "Texture filter set to %s")
 GS_DEFINE(OSD_TEXTURE_FILTER_NN, "nearest-neighbor")
 GS_DEFINE(OSD_TEXTURE_FILTER_BILINEAR, "bilinear")
